+++ tests/test_solution.py
@@ -1,5 +1,4 @@
-from pyscipopt import Model
-
+from pyscipopt import Model, SCIP_PARAMSETTING
 
 def test_solution_getbest():
     m = Model()
@@ -62,7 +61,29 @@ def test_solution_evaluation():
     assert sol[expr] == m.getVal(expr)
     assert sol[expr2] == m.getVal(expr2)
 
+def test_getPrimalRayVal():
+    m = Model()
+    x = m.addVar()
+    m.setObjective(x, "maximize")
+    m.setPresolve(SCIP_PARAMSETTING.OFF)
+    
+    m.hideOutput()
+    m.optimize()
+    
+    assert m.getPrimalRayVal(x) == 1
+    
+def test_getPrimalRay():
+    m = Model()
+    x = m.addVar()
+    y = m.addVar()
+    m.setObjective(x, "maximize")
+    m.setPresolve(SCIP_PARAMSETTING.OFF)
+    
+    m.hideOutput()
+    m.optimize()
+
+    assert m.getPrimalRay() == [1,0]
+
 
 if __name__ == "__main__":
-    test_solution_getbest()
-    test_solution_create()
+    test_getPrimalRayVal()